# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SistemaJava • Mercadinho PDV",
    "description": "Obtenha CSRF em /api/v1/auth/csrf; envie X-XSRF-TOKEN em mutações. Login retorna Bearer JWT.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/api/v1/products/{id}": {
      "get": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "product",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductView"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/lots/{id}": {
      "put": {
        "tags": [
          "stock-controller"
        ],
        "operationId": "editLot",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LotEdit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/writeoffs": {
      "get": {
        "tags": [
          "stock-controller"
        ],
        "operationId": "writeoffs",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageWriteoff"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "stock-controller"
        ],
        "operationId": "writeoff",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WriteoffInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Writeoff"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/products": {
      "get": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "products",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 40
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageProductView"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/lots": {
      "get": {
        "tags": [
          "stock-controller"
        ],
        "operationId": "lots",
        "parameters": [
          {
            "name": "productId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": {}
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "stock-controller"
        ],
        "operationId": "lot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LotInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer",
                    "format": "int64"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/writeoffs/summary": {
      "get": {
        "tags": [
          "stock-controller"
        ],
        "operationId": "summary",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/products/{id}/price-history": {
      "get": {
        "tags": [
          "stock-controller"
        ],
        "operationId": "priceHistory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": {}
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/inventory-counts": {
      "get": {
        "tags": [
          "stock-controller"
        ],
        "operationId": "inventory",
        "parameters": [
          {
            "name": "productId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": {}
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/cash/sessions/{id}": {
      "get": {
        "tags": [
          "cash-controller"
        ],
        "operationId": "summary_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Summary"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/csrf": {
      "get": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "csrf",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CsrfToken"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProductInput": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 0
          },
          "barcode": {
            "type": "string",
            "maxLength": 50,
            "minLength": 0
          },
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 0
          },
          "categoryId": {
            "type": "integer",
            "format": "int64"
          },
          "unit": {
            "type": "string",
            "pattern": "UN|KG"
          },
          "price": {
            "type": "number",
            "minimum": 0
          },
          "cost": {
            "type": "number",
            "minimum": 0
          },
          "minimumStock": {
            "type": "number",
            "minimum": 0
          },
          "expiresOn": {
            "type": "string",
            "format": "date"
          },
          "perishable": {
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "cost",
          "minimumStock",
          "name",
          "price",
          "unit"
        ]
      },
      "ProductView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "barcode": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "categoryId": {
            "type": "integer",
            "format": "int64"
          },
          "categoryName": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "cost": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          },
          "minimumStock": {
            "type": "number"
          },
          "expiresOn": {
            "type": "string",
            "format": "date"
          },
          "perishable": {
            "type": "boolean"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "LotEdit": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number",
            "minimum": 0
          },
          "expiresOn": {
            "type": "string",
            "format": "date"
          },
          "note": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          }
        },
        "required": [
          "quantity"
        ]
      },
...
      "WriteoffInput": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "pattern": "LOSS|INTERNAL"
          },
          "reason": {
            "type": "string",
            "maxLength": 80,
            "minLength": 0
          },
          "note": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WriteoffItemInput"
            },
            "maxItems": 100,
            "minItems": 0
          }
        },
        "required": [
          "items",
          "kind",
          "reason"
        ]
      },
      "WriteoffItemInput": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int64"
          },
          "quantity": {
            "type": "number",
            "minimum": 0
          },
          "lotId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "productId",
          "quantity"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "Writeoff": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "kind": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "totalCost": {
            "type": "number"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WriteoffItem"
            }
          }
        }
      },
      "WriteoffItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "productName": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "unitCost": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "productId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "LotInput": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int64"
          },
          "quantity": {
            "type": "number",
            "minimum": 0
          },
          "expiresOn": {
            "type": "string",
            "format": "date"
          },
          "note": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          }
        },
        "required": [
          "productId",
          "quantity"
        ]
      },
...
      "CashMovement": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "type": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "reason": {
            "type": "string"
          }
        }
      },
      "CashSession": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "openedAt": {
            "type": "string",
            "format": "date-time"
          },
          "openingAmount": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "closedAt": {
            "type": "string",
            "format": "date-time"
          },
          "closedBy": {
            "$ref": "#/components/schemas/User"
          },
          "expectedCash": {
            "type": "number"
          },
          "countedCash": {
            "type": "number"
          },
          "difference": {
            "type": "number"
          },
          "notes": {
            "type": "string"
          }
        }
      },
      "PaymentTotal": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string"
          },
          "total": {
            "type": "number"
          }
        }
      },
      "Summary": {
        "type": "object",
        "properties": {
          "session": {
            "$ref": "#/components/schemas/CashSession"
          },
          "sales": {
            "type": "integer",
            "format": "int64"
          },
          "cancelled": {
            "type": "integer",
            "format": "int64"
          },
          "salesTotal": {
            "type": "number"
          },
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentTotal"
            }
          },
          "cashSales": {
            "type": "number"
          },
          "creditReceipts": {
            "type": "number"
          },
          "supplies": {
            "type": "number"
          },
          "withdrawals": {
            "type": "number"
          },
          "expectedCash": {
            "type": "number"
          },
          "movements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CashMovement"
            }
          }
        }
      },
...
      "Login": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 80,
            "minLength": 0
          },
          "password": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          }
        },
        "required": [
          "password",
          "username"
        ]
      },
      "PageWriteoff": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "last": {
            "type": "boolean"
          },
          "first": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Writeoff"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "paged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "sorted": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      },
...
      "PageProductView": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "last": {
            "type": "boolean"
          },
          "first": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductView"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
      "CsrfToken": {
        "type": "object",
        "properties": {
          "headerName": {
            "type": "string"
          },
          "parameterName": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}